# specs/openapi/master-data.v1.yaml
openapi: 3.1.0
info:
  title: HIMS Master Data API
  version: 1.0.0
  description: 'HIMS Master Data catalog API — platform reference data (modules first; more surfaces phased in). Base path
    `/api/v1/master-data`. DELETE on modules is a recursive soft-delete (`is_deleted = true`). Gateway-level authentication
    (API keys / JWT) is expected before production; this spec does not model per-route security yet.

    '
servers:
- url: /api/v1/master-data
paths:
...
  /visitpad/diagnoses:
    get:
      tags:
      - Visitpad — Diagnoses
      summary: List diagnoses
      operationId: get_diagnoses_api_v1_master_data_visitpad_diagnoses_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadDiagnosisCategory'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadDiagnosisListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Visitpad — Diagnoses
      summary: Create diagnosis
      operationId: post_diagnosis_api_v1_master_data_visitpad_diagnoses_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadDiagnosisCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadDiagnosisSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /visitpad/diagnoses/{diagnosis_id}:
    get:
      tags:
      - Visitpad — Diagnoses
      summary: Get diagnosis
      operationId: get_diagnosis_api_v1_master_data_visitpad_diagnoses__diagnosis_id__get
      parameters:
      - name: diagnosis_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Diagnosis Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadDiagnosisSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Visitpad — Diagnoses
      summary: Update diagnosis
      operationId: patch_diagnosis_api_v1_master_data_visitpad_diagnoses__diagnosis_id__patch
      parameters:
      - name: diagnosis_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Diagnosis Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadDiagnosisUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadDiagnosisSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Visitpad — Diagnoses
      summary: Soft-delete diagnosis
      operationId: delete_diagnosis_api_v1_master_data_visitpad_diagnoses__diagnosis_id__delete
      parameters:
      - name: diagnosis_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Diagnosis Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadDiagnosisSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /visitpad/medicines:
    get:
      tags:
      - Visitpad — Medicines
      summary: List medicines
      operationId: get_medicines_api_v1_master_data_visitpad_medicines_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: schedule
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadMedicineSchedule'
          - type: 'null'
          title: Schedule
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadMedicineListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Visitpad — Medicines
      summary: Create medicine
      operationId: post_medicine_api_v1_master_data_visitpad_medicines_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadMedicineCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadMedicineSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /visitpad/medicines/{medicine_id}:
    get:
      tags:
      - Visitpad — Medicines
      summary: Get medicine
      operationId: get_medicine_api_v1_master_data_visitpad_medicines__medicine_id__get
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Medicine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadMedicineSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Visitpad — Medicines
      summary: Update medicine
      operationId: patch_medicine_api_v1_master_data_visitpad_medicines__medicine_id__patch
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Medicine Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadMedicineUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadMedicineSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Visitpad — Medicines
      summary: Soft-delete medicine
      operationId: delete_medicine_api_v1_master_data_visitpad_medicines__medicine_id__delete
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Medicine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadMedicineSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VisitpadDiagnosisCategory:
      type: string
      enum:
      - general
      - infectious
      - neoplastic
      - metabolic
      - psychiatric
      - injury
      - other
      title: VisitpadDiagnosisCategory
    VisitpadDiagnosisCreate:
      properties:
        icd10_code:
          type: string
          maxLength: 16
          minLength: 1
          title: Icd10 Code
        icd_version:
          $ref: '#/components/schemas/VisitpadIcdVersion'
        official_descriptor:
          type: string
          maxLength: 512
          minLength: 1
          title: Official Descriptor
        display_name:
          type: string
          maxLength: 512
          minLength: 1
          title: Display Name
        category:
          $ref: '#/components/schemas/VisitpadDiagnosisCategory'
        is_chronic_flag:
          type: boolean
          title: Is Chronic Flag
          default: false
        is_notifiable:
          type: boolean
          title: Is Notifiable
          default: false
        display_order:
          type: integer
          title: Display Order
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
      additionalProperties: false
      type: object
      required:
      - icd10_code
      - icd_version
      - official_descriptor
      - display_name
      - category
      title: VisitpadDiagnosisCreate
    VisitpadDiagnosisListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/VisitpadDiagnosisResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
      type: object
      required:
      - data
      - total
      title: VisitpadDiagnosisListResponse
    VisitpadDiagnosisResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        icd10_code:
          type: string
          title: Icd10 Code
        icd_version:
          $ref: '#/components/schemas/VisitpadIcdVersion'
        official_descriptor:
          type: string
          title: Official Descriptor
        display_name:
          type: string
          title: Display Name
        category:
          $ref: '#/components/schemas/VisitpadDiagnosisCategory'
        is_chronic_flag:
          type: boolean
          title: Is Chronic Flag
        is_notifiable:
          type: boolean
          title: Is Notifiable
        display_order:
          type: integer
          title: Display Order
        is_active:
          type: boolean
          title: Is Active
        is_deleted:
          type: boolean
          title: Is Deleted
        snomed_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Snomed Code
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - tenant_id
      - icd10_code
      - icd_version
      - official_descriptor
      - display_name
      - category
      - is_chronic_flag
      - is_notifiable
      - display_order
      - is_active
      - is_deleted
      - created_at
      - updated_at
      title: VisitpadDiagnosisResponse
    VisitpadDiagnosisSingleResponse:
      properties:
        data:
          $ref: '#/components/schemas/VisitpadDiagnosisResponse'
      type: object
      required:
      - data
      title: VisitpadDiagnosisSingleResponse
    VisitpadDiagnosisUpdate:
      properties:
        icd10_code:
          anyOf:
          - type: string
            maxLength: 16
            minLength: 1
          - type: 'null'
          title: Icd10 Code
        icd_version:
          anyOf:
          - $ref: '#/components/schemas/VisitpadIcdVersion'
          - type: 'null'
        official_descriptor:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Official Descriptor
        display_name:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Display Name
        category:
          anyOf:
          - $ref: '#/components/schemas/VisitpadDiagnosisCategory'
          - type: 'null'
        is_chronic_flag:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Chronic Flag
        is_notifiable:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Notifiable
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
        is_deleted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Deleted
      additionalProperties: false
      type: object
      title: VisitpadDiagnosisUpdate
    VisitpadIcdVersion:
      type: string
      enum:
      - ICD-10
      - ICD-11
      title: VisitpadIcdVersion
    VisitpadMedicineCreate:
      properties:
        code:
          type: string
          maxLength: 64
          minLength: 1
          title: Code
        display_name:
          type: string
          maxLength: 512
          minLength: 1
          title: Display Name
        generic_name:
          type: string
          maxLength: 512
          minLength: 1
          title: Generic Name
        short_name:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Short Name
        brand_names:
          items:
            type: string
          type: array
          maxItems: 50
          title: Brand Names
        drug_class:
          type: string
          maxLength: 256
          minLength: 1
          title: Drug Class
        drug_subclass:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Drug Subclass
        dosage_form:
          type: string
          maxLength: 128
          minLength: 1
          title: Dosage Form
        route_of_admin:
          items:
            type: string
          type: array
          maxItems: 20
          title: Route Of Admin
        strength_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Strength Value
        strength_unit:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Strength Unit
        strength_display:
          type: string
          maxLength: 256
          title: Strength Display
          default: ''
        concentration_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Concentration Value
        concentration_unit:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Concentration Unit
        volume_per_unit:
          anyOf:
          - type: number
          - type: 'null'
          title: Volume Per Unit
        sku_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Sku Code
        barcode:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Barcode
        pack_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pack Size
        pack_unit:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Pack Unit
        manufacturer:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Manufacturer
        storage_condition:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Storage Condition
        expiry_tracking:
          type: boolean
          title: Expiry Tracking
          default: false
        is_dispensable:
          type: boolean
          title: Is Dispensable
          default: true
        schedule:
          $ref: '#/components/schemas/VisitpadMedicineSchedule'
        is_controlled_substance:
          type: boolean
          title: Is Controlled Substance
          default: false
        is_narcotic:
          type: boolean
          title: Is Narcotic
          default: false
        requires_prescription:
          type: boolean
          title: Requires Prescription
          default: false
        is_restricted_antibiotic:
          type: boolean
          title: Is Restricted Antibiotic
          default: false
        allergen_classes:
          items:
            type: string
          type: array
          maxItems: 50
          title: Allergen Classes
        contraindications:
          items:
            type: string
          type: array
          maxItems: 50
          title: Contraindications
        search_tags:
          items:
            type: string
          type: array
          maxItems: 50
          title: Search Tags
        atc_code:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Atc Code
        rxnorm_code:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Rxnorm Code
        snomed_substance_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Substance Code
        snomed_product_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Product Code
        pregnancy_category:
          anyOf:
          - type: string
            maxLength: 8
          - type: 'null'
          title: Pregnancy Category
        lactation_safety:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Lactation Safety
        pediatric_use:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Pediatric Use
        max_dose_per_day_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Dose Per Day Value
        max_dose_per_day_unit:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Max Dose Per Day Unit
        black_box_warning:
          type: boolean
          title: Black Box Warning
          default: false
        black_box_warning_text:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Black Box Warning Text
        default_dose_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Default Dose Value
        default_dose_unit:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Default Dose Unit
        default_frequency:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Default Frequency
        default_duration_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Default Duration Days
        default_route:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Default Route
        default_instructions:
          anyOf:
          - type: string
            maxLength: 1024
          - type: 'null'
          title: Default Instructions
        typical_quantity:
          anyOf:
          - type: number
          - type: 'null'
          title: Typical Quantity
        notes:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Notes
        display_order:
          type: integer
          title: Display Order
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
      additionalProperties: false
      type: object
      required:
      - code
      - display_name
      - generic_name
      - drug_class
      - dosage_form
      - schedule
      title: VisitpadMedicineCreate
    VisitpadMedicineListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/VisitpadMedicineResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
      type: object
      required:
      - data
      - total
      title: VisitpadMedicineListResponse
...
    VisitpadMedicineSchedule:
      type: string
      enum:
      - otc
      - h
      - h1
      - s
      - x
      - unscheduled
      title: VisitpadMedicineSchedule
    VisitpadMedicineSingleResponse:
      properties:
        data:
          $ref: '#/components/schemas/VisitpadMedicineResponse'
      type: object
      required:
      - data
      title: VisitpadMedicineSingleResponse
...